samples: dfu: Move BLE MCUmgr firmware loader
moves from fw_loader/ble_mcumgr to single_slot sample
