+++ .github/test-spec.yml
@@ -126,7 +126,6 @@
   - "samples/zephyr/smp_svr_mini_boot/**/*"
   - "scripts/bootloader/**/*"
   - "samples/dfu/**/*"
-  - "samples/mcuboot/**/*"
   - "scripts/nrf_compress/**/*"
   - "scripts/west_commands/ncs*"
   - "subsys/bootloader/**/*"
@@ -155,7 +154,6 @@
   - "samples/zephyr/smp_svr_mini_boot/**/*"
   - "samples/bootloader/**/*"
   - "samples/dfu/**/*"
-  - "samples/mcuboot/**/*"
   - "scripts/bootloader/**/*"
   - "scripts/nrf_compress/**/*"
   - "scripts/west_commands/ncs*"
