+++ samples/dfu/single_slot/CMakeLists.txt
@@ -10,3 +10,4 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(single_slot)
 
 target_sources(app PRIVATE src/main.c)
+target_sources_ifdef(CONFIG_MCUMGR_TRANSPORT_BT app PRIVATE src/ble_smp.c)
