+++ samples/dfu/single_slot/boards/nrf52840dk_nrf52840.overlay
