+++ samples/dfu/single_slot/prj_ble_enter.conf
@@ -4,6 +4,7 @@ CONFIG_ZCBOR=y
 CONFIG_MCUMGR=y
 CONFIG_BT=y
 CONFIG_BT_PERIPHERAL=y
+CONFIG_BT_DEVICE_NAME="single_slot"
 
 # Allow for large Bluetooth data packets.
 CONFIG_BT_L2CAP_TX_MTU=498
@@ -23,7 +24,8 @@ CONFIG_MCUMGR_TRANSPORT_NETBUF_SIZE=2475
 CONFIG_MCUMGR_GRP_OS_MCUMGR_PARAMS=y
 CONFIG_MCUMGR_TRANSPORT_WORKQUEUE_STACK_SIZE=4608
 
-# Enable OS management with bootloader information and boot mode support
+# Enable OS management with bootloader information and boot mode support. The reset command with
+# the boot mode parameter is what requests entrance into the firmware loader.
 CONFIG_MCUMGR_GRP_OS=y
 CONFIG_MCUMGR_GRP_OS_RESET_BOOT_MODE=y
 CONFIG_MCUMGR_GRP_OS_BOOTLOADER_INFO=y
