+++ samples/dfu/single_slot/sample.yaml
@@ -19,6 +19,7 @@ tests:
       - dfu_single_slot
       - ci_samples_dfu
     platform_allow:
+      - nrf52840dk/nrf52840
       - nrf54l15dk/nrf54l15/cpuapp
       - nrf54h20dk/nrf54h20/cpuapp
       - nrf54lm20dk/nrf54lm20a/cpuapp
@@ -26,12 +27,31 @@ tests:
       - nrf54lm20dk/nrf54lm20b/cpuapp
       - nrf7120dk/nrf7120/cpuapp
     integration_platforms:
+      - nrf52840dk/nrf52840
       - nrf54l15dk/nrf54l15/cpuapp
       - nrf54h20dk/nrf54h20/cpuapp
       - nrf54lm20dk/nrf54lm20a/cpuapp
       - nrf54ls05dk/nrf54ls05b/cpuapp
       - nrf54lm20dk/nrf54lm20b/cpuapp
       - nrf7120dk/nrf7120/cpuapp
+  sample.dfu.single_slot.ble_enter:
+    tags:
+      - dfu_single_slot
+      - ci_samples_dfu
+      - bluetooth
+    platform_allow:
+      - nrf52840dk/nrf52840
+      - nrf54l15dk/nrf54l15/cpuapp
+    integration_platforms:
+      - nrf52840dk/nrf52840
+      - nrf54l15dk/nrf54l15/cpuapp
+    extra_args:
+      - FILE_SUFFIX=ble_enter
+    harness_config:
+      type: multi_line
+      regex:
+        - "Starting single_slot sample"
+        - "Advertising successfully started"
   sample.dfu.single_slot.usb:
     tags:
       - dfu_single_slot
