+++ samples/dfu/single_slot/src/ble_smp.c
@@ -1,0 +1,92 @@
+/*
+ * Copyright (c) 2026 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#include <zephyr/kernel.h>
+#include <zephyr/settings/settings.h>
+#include <zephyr/bluetooth/bluetooth.h>
+#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/conn.h>
+#include <zephyr/mgmt/mcumgr/transport/smp_bt.h>
+#include <zephyr/logging/log.h>
+
+#include "ble_smp.h"
+
+LOG_MODULE_REGISTER(single_slot_ble_smp, LOG_LEVEL_INF);
+
+static struct k_work adv_work;
+
+static const struct bt_data ad[] = {
+	BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
+	BT_DATA_BYTES(BT_DATA_UUID128_ALL, SMP_BT_SVC_UUID_VAL),
+};
+
+static const struct bt_data sd[] = {
+	BT_DATA(BT_DATA_NAME_COMPLETE, CONFIG_BT_DEVICE_NAME, (sizeof(CONFIG_BT_DEVICE_NAME) - 1)),
+};
+
+static void adv_work_handler(struct k_work *work)
+{
+	int rc = bt_le_adv_start(BT_LE_ADV_CONN_FAST_2, ad, ARRAY_SIZE(ad), sd, ARRAY_SIZE(sd));
+
+	if (rc) {
+		LOG_ERR("Advertising failed to start: %d", rc);
+		return;
+	}
+
+	LOG_INF("Advertising successfully started, device name: %s", CONFIG_BT_DEVICE_NAME);
+}
+
+static void advertising_start(void)
+{
+	k_work_submit(&adv_work);
+}
+
+static void connected(struct bt_conn *conn, uint8_t err)
+{
+	if (err) {
+		LOG_ERR("Connection failed: 0x%02x (%s)", err, bt_hci_err_to_str(err));
+		return;
+	}
+
+	LOG_INF("Connected");
+}
+
+static void disconnected(struct bt_conn *conn, uint8_t reason)
+{
+	LOG_INF("Disconnected, reason: 0x%02x (%s)", reason, bt_hci_err_to_str(reason));
+}
+
+static void recycled_cb(void)
+{
+	advertising_start();
+}
+
+BT_CONN_CB_DEFINE(conn_callbacks) = {
+	.connected    = connected,
+	.disconnected = disconnected,
+	.recycled     = recycled_cb,
+};
+
+int ble_smp_init(void)
+{
+	int rc = bt_enable(NULL);
+
+	if (rc) {
+		LOG_ERR("Bluetooth init failed: %d", rc);
+		return rc;
+	}
+
+	LOG_INF("Bluetooth initialized");
+
+	if (IS_ENABLED(CONFIG_SETTINGS)) {
+		settings_load();
+	}
+
+	k_work_init(&adv_work, adv_work_handler);
+	advertising_start();
+
+	return 0;
+}
