+++ samples/dfu/single_slot/sysbuild.conf
@@ -7,4 +7,3 @@
 SB_CONFIG_BOOTLOADER_MCUBOOT=y
 SB_CONFIG_MCUBOOT_MODE_FIRMWARE_UPDATER=y
 SB_CONFIG_FIRMWARE_LOADER_IMAGE_BLE_MCUMGR=y
-SB_CONFIG_BOOT_SIGNATURE_TYPE_PURE=y
