+++ samples/dfu/single_slot/sysbuild/ble_mcumgr/boards/nrf52840dk_nrf52840.overlay
