+++ samples/dfu/single_slot/sysbuild/mcuboot/boards/nrf52840dk_nrf52840.conf
