+++ samples/dfu/single_slot/sysbuild/mcuboot/boards/nrf52840dk_nrf52840.overlay
@@ -4,10 +4,14 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 
-#include "../../../boards/nrf54l15dk_nrf54l15_cpuapp.overlay"
+#include "../../../boards/nrf52840dk_nrf52840.overlay"
 
 / {
 	chosen {
 		zephyr,code-partition = &boot_partition;
+	};
+
+	aliases {
+		mcuboot-button0 = &button0;
 	};
 };
