+++ samples/dfu/single_slot/sysbuild_usb.conf
@@ -7,4 +7,3 @@
 SB_CONFIG_BOOTLOADER_MCUBOOT=y
 SB_CONFIG_MCUBOOT_MODE_FIRMWARE_UPDATER=y
 SB_CONFIG_FIRMWARE_LOADER_IMAGE_USB_MCUMGR=y
-SB_CONFIG_BOOT_SIGNATURE_TYPE_PURE=y
