+++ samples/dfu/single_slot/sysbuild_usb_enter.conf
@@ -7,6 +7,5 @@
 SB_CONFIG_BOOTLOADER_MCUBOOT=y
 SB_CONFIG_MCUBOOT_MODE_FIRMWARE_UPDATER=y
 SB_CONFIG_FIRMWARE_LOADER_IMAGE_USB_MCUMGR=y
-SB_CONFIG_BOOT_SIGNATURE_TYPE_PURE=y
 
 SB_CONFIG_MCUBOOT_MODE_FIRMWARE_UPDATER_BOOT_MODE_ENTRANCE=y
