--- samples/mcuboot/firmware_loader_entrance/CMakeLists.txt
@@ -1,12 +1,0 @@
-#
-# Copyright (c) 2025 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-cmake_minimum_required(VERSION 3.20.0)
-
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(firmware_loader_entrance)
-
-target_sources(app PRIVATE src/main.c)
