--- samples/mcuboot/firmware_loader_entrance/Kconfig.sysbuild
@@ -1,10 +1,0 @@
-#
-# Copyright (c) 2026 Nordic Semiconductor
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-config PARTITION_MANAGER
-	default n
-
-source "${ZEPHYR_BASE}/share/sysbuild/Kconfig"
