--- samples/mcuboot/firmware_loader_entrance/sample.yaml
@@ -1,21 +1,0 @@
-sample:
-  name: Firmware loader entrance
-  description: Demo application for entering the firmware loader application
-common:
-  sysbuild: true
-tests:
-  sample.mcuboot.firmware_loader_entrance:
-    harness: console
-    harness_config:
-      type: multi_line
-      regex:
-        - "Advertising successfully started"
-    tags:
-      - bluetooth
-      - ci_samples_mcuboot
-    platform_allow:
-      - nrf52840dk/nrf52840
-      - nrf54l15dk/nrf54l15/cpuapp
-    integration_platforms:
-      - nrf52840dk/nrf52840
-      - nrf54l15dk/nrf54l15/cpuapp
