--- samples/mcuboot/firmware_loader_entrance/src/main.c
@@ -1,183 +1,0 @@
-/*
- * Copyright (c) 2025 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
- */
-
-#include <zephyr/kernel.h>
-#include <zephyr/settings/settings.h>
-#include <zephyr/bluetooth/bluetooth.h>
-#include <zephyr/bluetooth/hci.h>
-#include <zephyr/bluetooth/conn.h>
-#include <zephyr/bluetooth/uuid.h>
-#include <zephyr/bluetooth/gatt.h>
-#include <zephyr/mgmt/mcumgr/transport/smp_bt.h>
-#include <zephyr/logging/log.h>
-
-LOG_MODULE_REGISTER(firmware_entrance_loader, LOG_LEVEL_DBG);
-
-static struct k_work adv_work;
-
-static const struct bt_data ad[] = {
-	BT_DATA_BYTES(BT_DATA_FLAGS, (BT_LE_AD_GENERAL | BT_LE_AD_NO_BREDR)),
-	BT_DATA_BYTES(BT_DATA_UUID128_ALL, SMP_BT_SVC_UUID_VAL),
-};
-
-static const struct bt_data sd[] = {
-	BT_DATA(BT_DATA_NAME_COMPLETE, CONFIG_BT_DEVICE_NAME, (sizeof(CONFIG_BT_DEVICE_NAME) - 1)),
-};
-
-static void adv_work_handler(struct k_work *work)
-{
-	int rc = bt_le_adv_start(BT_LE_ADV_CONN_FAST_2, ad, ARRAY_SIZE(ad), sd, ARRAY_SIZE(sd));
-
-	if (rc) {
-		LOG_ERR("Advertising failed to start: %d", rc);
-		return;
-	}
-
-	LOG_INF("Advertising successfully started, device name: %s", CONFIG_BT_DEVICE_NAME);
-}
-
-static void advertising_start(void)
-{
-	k_work_submit(&adv_work);
-}
-
-static void connected(struct bt_conn *conn, uint8_t err)
-{
-	if (err) {
-		LOG_ERR("Connection failed: 0x%02x (%s)", err, bt_hci_err_to_str(err));
-		return;
-	}
-
-	LOG_INF("Connected");
-}
-
-static void disconnected(struct bt_conn *conn, uint8_t reason)
-{
-	LOG_INF("Disconnected, reason: 0x%02x (%s)", reason, bt_hci_err_to_str(reason));
-}
-
-static void recycled_cb(void)
-{
-	LOG_INF("Connection object available from previous conn. Disconnect is complete!");
-	advertising_start();
-}
-
-#ifdef CONFIG_BT_SMP
-static void security_changed(struct bt_conn *conn, bt_security_t level,
-			     enum bt_security_err err)
-{
-	char addr[BT_ADDR_LE_STR_LEN];
-
-	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
-
-	if (!err) {
-		LOG_INF("Security changed: %s level %u", addr, level);
-	} else {
-		LOG_INF("Security failed: %s level %u error: %d (%s)", addr, level, err,
-			bt_security_err_to_str(err));
-	}
-}
-#endif
-
-BT_CONN_CB_DEFINE(conn_callbacks) = {
-	.connected        = connected,
-	.disconnected     = disconnected,
-	.recycled         = recycled_cb,
-#ifdef CONFIG_BT_SMP
-	.security_changed = security_changed,
-#endif
-};
-
-#if defined(CONFIG_BT_SMP)
-static void auth_passkey_display(struct bt_conn *conn, unsigned int passkey)
-{
-	char addr[BT_ADDR_LE_STR_LEN];
-
-	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
-
-	LOG_INF("Passkey for %s: %06u", addr, passkey);
-}
-
-static void auth_cancel(struct bt_conn *conn)
-{
-	char addr[BT_ADDR_LE_STR_LEN];
-
-	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
-
-	LOG_INF("Pairing cancelled: %s", addr);
-}
-
-static void pairing_complete(struct bt_conn *conn, bool bonded)
-{
-	char addr[BT_ADDR_LE_STR_LEN];
-
-	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
-
-	LOG_INF("Pairing completed: %s, bonded: %d", addr, bonded);
-}
-
-static void pairing_failed(struct bt_conn *conn, enum bt_security_err reason)
-{
-	char addr[BT_ADDR_LE_STR_LEN];
-
-	bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
-
-	LOG_INF("Pairing failed conn: %s, reason: %d (%s)", addr, reason,
-		bt_security_err_to_str(reason));
-}
-
-static struct bt_conn_auth_cb conn_auth_callbacks = {
-	.passkey_display = auth_passkey_display,
-	.cancel = auth_cancel,
-};
-
-static struct bt_conn_auth_info_cb conn_auth_info_callbacks = {
-	.pairing_complete = pairing_complete,
-	.pairing_failed = pairing_failed
-};
-#else
-static struct bt_conn_auth_cb conn_auth_callbacks;
-static struct bt_conn_auth_info_cb conn_auth_info_callbacks;
-#endif
-
-int main(void)
-{
-	int rc;
-
-	if (IS_ENABLED(CONFIG_BT_SMP)) {
-		rc = bt_conn_auth_cb_register(&conn_auth_callbacks);
-
-		if (rc) {
-			LOG_ERR("Failed to register authorization callbacks.");
-			return 0;
-		}
-
-		rc = bt_conn_auth_info_cb_register(&conn_auth_info_callbacks);
-
-		if (rc) {
-			LOG_ERR("Failed to register authorization info callbacks.");
-			return 0;
-		}
-	}
-
-	rc = bt_enable(NULL);
-
-	if (rc) {
-		LOG_ERR("Bluetooth init failed: %d", rc);
-		return 0;
-	}
-
-	LOG_INF("Bluetooth initialized");
-
-	if (IS_ENABLED(CONFIG_SETTINGS)) {
-		settings_load();
-	}
-
-	k_work_init(&adv_work, adv_work_handler);
-	advertising_start();
-
-	return 0;
-}
