--- samples/mcuboot/firmware_loader_entrance/sysbuild/ble_mcumgr/boards/nrf54l15dk_nrf54l15_cpuapp.overlay
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2026 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
- */
-
-#include "../../../boards/nrf54l15dk_nrf54l15_cpuapp.overlay"
