--- samples/mcuboot/firmware_loader_entrance/sysbuild/ble_mcumgr/prj.conf
@@ -1,126 +1,0 @@
-#
-# Copyright (c) 2026 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-
-# Enable MCUmgr and dependencies.
-CONFIG_NET_BUF=y
-CONFIG_ZCBOR=y
-CONFIG_CRC=y
-CONFIG_MCUMGR=y
-CONFIG_STREAM_FLASH=y
-CONFIG_FLASH_MAP=y
-
-# Some command handlers require a large stack.
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2304
-CONFIG_MAIN_STACK_SIZE=2176
-
-# Ensure an MCUboot-compatible binary is generated.
-CONFIG_BOOTLOADER_MCUBOOT=y
-
-# Enable flash operations.
-CONFIG_FLASH=y
-
-# Enable most core commands.
-CONFIG_IMG_MANAGER=y
-CONFIG_MCUMGR_GRP_IMG=y
-CONFIG_MCUMGR_GRP_OS=y
-CONFIG_MCUMGR_GRP_OS_BOOTLOADER_INFO=y
-CONFIG_MCUMGR_GRP_OS_ECHO=n
-CONFIG_MCUMGR_GRP_OS_INFO=n
-
-CONFIG_BT=y
-CONFIG_BT_PERIPHERAL=y
-
-# Allow for large Bluetooth data packets.
-CONFIG_BT_L2CAP_TX_MTU=247
-CONFIG_BT_BUF_ACL_RX_SIZE=251
-CONFIG_BT_BUF_ACL_TX_SIZE=251
-
-# Enable the Bluetooth mcumgr transport (unauthenticated).
-CONFIG_MCUMGR_TRANSPORT_BT=y
-CONFIG_MCUMGR_TRANSPORT_BT_CONN_PARAM_CONTROL=y
-
-# Enable the Shell mcumgr transport.
-CONFIG_BASE64=y
-
-# Enable the mcumgr Packet Reassembly feature over Bluetooth and its configuration dependencies.
-# MCUmgr buffer size is optimized to fit one SMP packet divided into five Bluetooth Write Commands,
-# transmitted with the maximum possible MTU value: 247 bytes.
-CONFIG_MCUMGR_TRANSPORT_BT_REASSEMBLY=y
-CONFIG_MCUMGR_TRANSPORT_NETBUF_SIZE=1220
-CONFIG_MCUMGR_GRP_OS_MCUMGR_PARAMS=y
-CONFIG_MCUMGR_TRANSPORT_WORKQUEUE_STACK_SIZE=4608
-CONFIG_BT_DEVICE_NAME="FW loader"
-
-
-# Enable the storage erase command.
-CONFIG_MCUMGR_GRP_ZBASIC=y
-CONFIG_MCUMGR_GRP_ZBASIC_STORAGE_ERASE=y
-
-# Disable Bluetooth ping support
-CONFIG_BT_CTLR_LE_PING=n
-
-# Disable shell commands that are not needed
-CONFIG_CLOCK_CONTROL_NRF_SHELL=n
-CONFIG_DEVICE_SHELL=n
-CONFIG_DEVMEM_SHELL=n
-CONFIG_FLASH_SHELL=n
-########
-
-# Disable unneeded drivers, peripherals and features to optimize for size
-CONFIG_SIZE_OPTIMIZATIONS=y
-
-# Enable LTO
-CONFIG_LTO=y
-CONFIG_ISR_TABLES_LOCAL_DECLARATION=y
-
-CONFIG_I2C=n
-CONFIG_WATCHDOG=n
-CONFIG_SPI_NOR=n
-CONFIG_GPIO=n
-
-CONFIG_FPU=n
-
-CONFIG_BOOT_BANNER=n
-CONFIG_NCS_BOOT_BANNER=n
-CONFIG_BOOT_DELAY=0
-CONFIG_PRINTK=n
-
-CONFIG_ARM_MPU=n
-
-CONFIG_TIMESLICING=n
-CONFIG_COMMON_LIBC_MALLOC=y
-CONFIG_LOG=n
-CONFIG_SERIAL=n
-CONFIG_UART_CONSOLE=n
-CONFIG_ASSERT=n
-
-CONFIG_BT_HCI_VS=n
-
-CONFIG_BT_DEBUG_NONE=y
-CONFIG_BT_ASSERT=n
-CONFIG_BT_DATA_LEN_UPDATE=n
-CONFIG_BT_PHY_UPDATE=n
-CONFIG_BT_GATT_CACHING=n
-
-CONFIG_BT_GATT_ENFORCE_SUBSCRIPTION=n
-CONFIG_BT_GATT_READ_MULTIPLE=n
-CONFIG_BT_GATT_READ_MULT_VAR_LEN=n
-CONFIG_BT_GAP_AUTO_UPDATE_CONN_PARAMS=n
-
-CONFIG_BT_CTLR_PRIVACY=n
-CONFIG_BT_CTLR_PHY_2M=n
-
-CONFIG_PSA_USE_CRACEN_CIPHER_DRIVER=n
-CONFIG_PSA_USE_CRACEN_AEAD_DRIVER=n
-CONFIG_PSA_USE_CRACEN_MAC_DRIVER=n
-CONFIG_PSA_USE_CRACEN_KEY_AGREEMENT_DRIVER=n
-CONFIG_PSA_USE_CRACEN_ASYMMETRIC_ENCRYPTION_DRIVER=n
-CONFIG_PSA_USE_CRACEN_KEY_MANAGEMENT_DRIVER=n
-CONFIG_PSA_USE_CRACEN_KEY_DERIVATION_DRIVER=n
-CONFIG_PSA_USE_CRACEN_PAKE_DRIVER=n
-CONFIG_PSA_USE_CRACEN_ASYMMETRIC_SIGNATURE_DRIVER=n
-CONFIG_PSA_USE_CRACEN_HASH_DRIVER=n
-
-CONFIG_BT_GATT_SERVICE_CHANGED=n
