--- samples/mcuboot/firmware_loader_entrance/sysbuild/mcuboot.conf
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2025 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-CONFIG_BOOT_FIRMWARE_LOADER_ENTRANCE_GPIO=y
-CONFIG_BOOT_FIRMWARE_LOADER_NO_APPLICATION=y
