--- samples/mcuboot/firmware_loader_entrance/sysbuild/mcuboot/boards/nrf52840dk_nrf52840.overlay
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2026 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
- */
-
-#include "../../../boards/nrf52840dk_nrf52840.overlay"
-
-/ {
-	chosen {
-		zephyr,code-partition = &boot_partition;
-	};
-};
