--- samples/mcuboot/firmware_loader_entrance/sysbuild/mcuboot/boards/nrf54l15dk_nrf54l15_cpuapp.conf
@@ -1,20 +1,0 @@
-#
-# Copyright (c) 2026 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-
-# Configuration copied from mcuboot/boot/zephyr/socs/nrf54l15_cpuapp.conf
-
-CONFIG_BOOT_MAX_IMG_SECTORS=256
-
-# Ensure that the SPI NOR driver is disabled by default
-CONFIG_SPI_NOR=n
-
-CONFIG_BOOT_WATCHDOG_FEED=n
-
-# Ensure the fastest RRAM write operations
-CONFIG_NRF_RRAM_WRITE_BUFFER_SIZE=32
-
-# Link Time Optimizations
-CONFIG_ISR_TABLES_LOCAL_DECLARATION=y
-CONFIG_LTO=y
