--- samples/mcuboot/firmware_loader_entrance/sysbuild/mcuboot/prj.conf
@@ -1,40 +1,0 @@
-#
-# Copyright (c) 2026 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-CONFIG_PM=n
-
-CONFIG_MAIN_STACK_SIZE=10240
-
-CONFIG_BOOT_SWAP_SAVE_ENCTLV=n
-CONFIG_BOOT_ENCRYPT_IMAGE=n
-
-CONFIG_BOOT_UPGRADE_ONLY=n
-CONFIG_BOOT_BOOTSTRAP=n
-
-### mbedTLS has its own heap
-# CONFIG_HEAP_MEM_POOL_SIZE is not set
-
-CONFIG_FLASH=y
-
-### Various Zephyr boards enable features that we don't want.
-# CONFIG_BT is not set
-# CONFIG_I2C is not set
-
-CONFIG_LOG=y
-CONFIG_LOG_MODE_MINIMAL=y
-### Ensure Zephyr logging changes don't use more resources
-CONFIG_LOG_DEFAULT_LEVEL=0
-### Use info log level by default
-CONFIG_MCUBOOT_LOG_LEVEL_INF=y
-### Decrease footprint by ~4 KB in comparison to CBPRINTF_COMPLETE=y
-CONFIG_CBPRINTF_NANO=y
-### Use picolibc to reduce flash usage
-CONFIG_PICOLIBC=y
-### Disable malloc arena because we don't need it
-CONFIG_COMMON_LIBC_MALLOC_ARENA_SIZE=0
-
-# NCS boot banner
-CONFIG_NCS_APPLICATION_BOOT_BANNER_STRING="MCUboot"
